Refactor registration schema
