--- fideslog/api/schemas/user_registration_event.py
@@ -1,53 +1,0 @@
-# pylint: disable= no-self-argument, no-self-use
-
-from datetime import datetime, timezone
-
-from pydantic import BaseModel, EmailStr, Field, validator
-
-
-class UserRegistrationEvent(BaseModel):
-    """The model for user registration events."""
-
-    analytics_id: str = Field(
-        ...,
-        description="The fully anonymized, globally unique identifier for the event sender.",
-    )
-    email: EmailStr = Field(
-        ...,
-        description="The user's email address.",
-    )
-    organization: str = Field(
-        ...,
-        description="The organization in which the user is registered.",
-    )
-    registered_at: datetime = Field(
-        ...,
-        description="The UTC timestamp when the user registration event occurred, in ISO 8601 format. Must include "
-        "the UTC timezone, and represent a datetime in the past.",
-    )
-
-    @validator("analytics_id")
-    def check_not_an_email_address(cls, value: str) -> str:
-        """
-        Validate that analytics_id does not contain an email address literal.
-        """
-
-        assert value.find("@") == -1, "analytics_id must not be identifiable"
-        return value
-
-    @validator("registered_at")
-    def check_in_the_past(cls, value: datetime) -> datetime:
-        """
-        Validate that the event creation timestamp is in the past.
-        """
-
-        assert (
-            value.tzinfo == timezone.utc
-        ), "registered_at must be an explicit UTC timestamp"
-        assert value < datetime.now(timezone.utc), "registered_at must be in the past"
-        return value
-
-    class Config:
-        """Modifies pydantic behavior."""
-
-        orm_mode = True
